requirements.txt
Django==3.2.9
Faker==13.6.0
Jinja2==3.1.1
MarkupSafe==2.1.1
Pillow==9.1.1
PyJWT==2.3.0
amqp==5.1.1
argon2-cffi==21.1.0
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==21.10b0
bs4==0.0.1
cachetools==4.2.4
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
cycler==0.11.0
defusedxml==0.7.1
django-allauth==0.45.0
django-anymail==8.4
django-celery-beat==2.2.1
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-decouple==2.1
django-environ==0.8.1
django-extensions==3.1.3
django-filter==21.1
django-model-utils==4.2.0
django-redis==5.0.0
django-timezone-field==4.2.3
djangorestframework-simplejwt==5.0.0
djangorestframework==3.12.4
djongo==1.3.6
drf-yasg==1.20.0
et-xmlfile==1.1.0
factory-boy==3.2.1
flower==1.0.0
fonttools==4.33.3
google-api-core==2.2.2
google-api-python-client==2.29.0
google-auth-httplib2==0.1.0
google-auth==2.3.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.20.4
humanize==4.0.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isodate==0.6.1
itypes==1.2.0
joblib==1.1.0
kiwisolver==1.4.2
kombu==5.2.4
matplotlib==3.5.2
mypy-extensions==0.4.3
mysqlclient==2.1.0
nltk==3.7
numpy==1.22.4
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psycopg2==2.9.1
py==1.11.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pymongo==3.12.1
pyparsing==3.0.8
pytest==7.1.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
redis==3.5.3
regex==2022.4.24
requests-oauthlib==1.3.1
requests==2.27.1
rsa==4.8
ruamel.yaml.clib==0.2.6
ruamel.yaml==0.17.21
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
sre-yield==1.2
text-unidecode==1.3
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
typing-extensions==3.10.0.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
